Reduce global state, allow operating w/ other tax add-ons
- VAT Manager only shows when necessary
- VAT only displays when unknown location or customer in EU ( made need setting? )
- Reduce reliance on session. Generates summarized info from cart at transaction.

This needs more effort to work with how VAT #'s work for non session based carts.
Looks like we need to introduce some form of light weight cart meta storage that
could be persisted to the session repo when necessary. But also be persisted to
transaction meta after purchase. That way a VAT # can be saved to cart meta.